If you're reading this in irc you should click on it to read the full list of changes. Did all the actor name changes such as from ZekeShotgun to MelZekeshotgun, and all needed code changes to accomodate these changes, as requested.
Added a CVar TSP_ShellPool. It defaults to true, which produces current normal Zeke Shotgun behavior. If set to false, the various Zeke Shotgun shell types do not share the same ammo pool and instead have max limits of 32/40 for buckshot and 16/24 for all other shell types, for a total shell carry capacity of 72/104.

Fixed idfa/idkfa/give all cheats, so that if TSP_ShellPool is set to true, the player doesn't end up with 96 of every shell type and instead has a mix of shells which total 96. Functions correctly with both TSP_ShellPool settings.

Fixed minor issue with character selection screen using the wrong cursor.